requirements.txt
Flask
openai
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
schedule
python-dotenv
gunicorn
